refactor: test 코드들을 하나의 폴더로 모으고, 도메인 별로 폴더 분리
