+++ frontend/src/__test__/article/useGetCategoryArticles.test.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BrowserRouter } from 'react-router-dom';
 
+import useGetCategoryArticles from '@/hooks/article/useGetCategoryArticles';
 import { ArticleHandler } from '@/mock';
-import useGetCategoryArticles from '@/pages/CategoryArticles/hooks/useGetCategoryArticles';
 import { theme } from '@/styles/Theme';
 import { ThemeProvider } from '@emotion/react';
 import { renderHook } from '@testing-library/react-hooks';
